Adds support for @Native and imports
