+++ templates/dsl/entities.hbs
@@ -1,3 +1,4 @@
+{{#kaplang-render this}}
 {{#switch type}}
     {{#case 'datatype'}}
         {{> dsl/entitybase.java}}
@@ -7,3 +8,4 @@
         {{> dsl/entitybase.java}}
     {{/case}}
 {{/switch}}
+{{/kaplang-render}}
